Password Generator:
generate_password: This function generates a random password using letters, numbers, and symbols. The generated password is copied to the clipboard and inserted into the password entry field.
Save Password:

save_password: This function saves the website, email/username, and password to a file. It checks if any fields are empty and displays a warning if they are. If all fields are filled, it asks for confirmation before saving the data to a file.